requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
pytesseract
spacy
pdf2image